Intended version
This version animates a triangle upon a straight line.
The speed is ultimately controlled by the frame rate of the browser. That is to say that the next position of the object doesn't reach in a predefined time as expected but increments per draw.

This would not be ideal for multiple objects or physics engines or both.