Intended version, proper timer
A proper timing method was used here.
This method prevents the draw call (and the state) from stacking and achieves the intended every 5milisecond call. When a set interval is used on a single threaded system the draw call can take longer than the interval set, and therefore will be called immediately. 
The intention here is to allow CPU and GPU cool down time and this is the proper way to achieve that.

The states animation loop ensures positions are in line with intention to hit target destination without knowing a developers mind or creating an API in order to achieve constant speed regardless of fluctuating hardware issues such as frame rate drops